+++ src/components/Title/Title.stories.tsx
@@ -7,12 +7,13 @@ import { Title as TitleComponent, TitleProps } from "./";
 export default {
   title: "Components/Title",
   component: TitleComponent,
-  argTypes: {
-    backgroundColor: { control: "color" },
-  },
 } as Meta;
 
-const Template: Story<TitleProps> = (args) => <TitleComponent {...args} />;
+const Template: Story<TitleProps> = (args) => (
+  <div style={{ backgroundColor: '#333333' }}>
+    <TitleComponent {...args} />
+  </div>
+)
 
 export const Title = Template.bind({});
 Title.args = {