+++ src/components/Title/index.tsx
@@ -1,7 +1,16 @@
+import styled from '@emotion/styled'
+
 export interface TitleProps {
   label: string;
 }
 
 export function Title({ label }: TitleProps) {
-  return <h1>{label}</h1>;
+  return <StyledTitle data-elm-id="title">{label}</StyledTitle>;
 }
+
+const StyledTitle = styled.h1`
+  font-family: Helvetica;
+  color: white;
+  padding: 16px;
+  text-align: center;
+`