+++ src/components/index.ts
@@ -1,1 +1,4 @@
 export * from "./Title";
+export * from "./Menu";
+export * from "./MenuItem";
+export * from "./PostContent";