+++ src/setupTests.ts
@@ -1,4 +1,4 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import Enzyme from 'enzyme';
-import ReactSixteenAdapter from 'enzyme-adapter-react-16';
-Enzyme.configure({ adapter: new ReactSixteenAdapter() });
+import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
+Enzyme.configure({ adapter: new Adapter() });